requirements.txt
discord.py
plexapi
requests
aiohttp
python-dateutil
uptime_kuma_api
python-dotenv